refactor(languages): remove completion related code
Closes #3300
