+++ src/__tests__/list/manager.test.ts
@@ -453,6 +453,7 @@ describe('list', () => {
       }
       manager.registerList(list)
       await manager.start(['test'])
+      await helper.wait(10)
       let msg = await helper.getCmdline()
       expect(msg).toMatch('test error')
     })
