+++ src/index.ts
@@ -7,7 +7,7 @@ import fetch from './model/fetch'
 import download from './model/download'
 import Highligher from './model/highligher'
 import services from './services'
-import sources from './sources'
+import sources from './sources/index'
 import workspace from './workspace'
 import window from './window'
 import extensions from './extensions'
