+++ src/languages.ts
@@ -1,9 +1,9 @@
-import { Neovim } from '@chemzqm/neovim'
-import { CallHierarchyIncomingCall, CallHierarchyItem, CallHierarchyOutgoingCall, CancellationToken, CancellationTokenSource, CodeAction, CodeActionContext, CodeActionKind, CodeLens, ColorInformation, ColorPresentation, CompletionItem, CompletionItemKind, CompletionList, CompletionTriggerKind, DefinitionLink, Disposable, DocumentHighlight, DocumentLink, DocumentSelector, DocumentSymbol, Emitter, Event, FoldingRange, FormattingOptions, Hover, InsertReplaceEdit, InsertTextFormat, LinkedEditingRanges, Location, LocationLink, Position, Range, SelectionRange, SemanticTokens, SemanticTokensDelta, SemanticTokensLegend, SignatureHelp, SignatureHelpContext, SymbolInformation, TextEdit, WorkspaceEdit } from 'vscode-languageserver-protocol'
+import { CallHierarchyIncomingCall, CallHierarchyItem, CallHierarchyOutgoingCall, CancellationToken, CancellationTokenSource, CodeAction, CodeActionContext, CodeActionKind, CodeLens, ColorInformation, ColorPresentation, DefinitionLink, Disposable, DocumentHighlight, DocumentLink, DocumentSelector, DocumentSymbol, Emitter, Event, FoldingRange, FormattingOptions, Hover, LinkedEditingRanges, Location, LocationLink, Position, Range, SelectionRange, SemanticTokens, SemanticTokensDelta, SemanticTokensLegend, SignatureHelp, SignatureHelpContext, SymbolInformation, TextEdit, WorkspaceEdit } from 'vscode-languageserver-protocol'
 import { TextDocument } from 'vscode-languageserver-textdocument'
-import commands from './commands'
+import DiagnosticCollection from './diagnostic/collection'
 import diagnosticManager from './diagnostic/manager'
 import { CallHierarchyProvider, CodeActionProvider, CodeLensProvider, CompletionItemProvider, DeclarationProvider, DefinitionProvider, DocumentColorProvider, DocumentFormattingEditProvider, DocumentHighlightProvider, DocumentLinkProvider, DocumentRangeFormattingEditProvider, DocumentRangeSemanticTokensProvider, DocumentSemanticTokensProvider, DocumentSymbolProvider, FoldingContext, FoldingRangeProvider, HoverProvider, ImplementationProvider, LinkedEditingRangeProvider, OnTypeFormattingEditProvider, ReferenceContext, ReferenceProvider, RenameProvider, SelectionRangeProvider, SignatureHelpProvider, TypeDefinitionProvider, WorkspaceSymbolProvider } from './provider'
+import CallHierarchyManager from './provider/callHierarchyManager'
 import CodeActionManager from './provider/codeActionManager'
 import CodeLensManager from './provider/codeLensManager'
 import DeclarationManager from './provider/declarationManager'
@@ -17,44 +17,17 @@ import FormatManager from './provider/formatManager'
 import FormatRangeManager from './provider/formatRangeManager'
 import HoverManager from './provider/hoverManager'
 import ImplementationManager from './provider/implementationManager'
+import LinkedEditingRangeManager from './provider/linkedEditingRangeManager'
 import OnTypeFormatManager from './provider/onTypeFormatManager'
-import SelectionRangeManager from './provider/selectionRangeManager'
 import ReferenceManager from './provider/referenceManager'
 import RenameManager from './provider/renameManager'
+import SelectionRangeManager from './provider/selectionRangeManager'
+import SemanticTokensManager from './provider/semanticTokensManager'
+import SemanticTokensRangeManager from './provider/semanticTokensRangeManager'
 import SignatureManager from './provider/signatureManager'
 import TypeDefinitionManager from './provider/typeDefinitionManager'
 import WorkspaceSymbolManager from './provider/workspaceSymbolsManager'
-import CallHierarchyManager from './provider/callHierarchyManager'
-import SemanticTokensManager from './provider/semanticTokensManager'
-import SemanticTokensRangeManager from './provider/semanticTokensRangeManager'
-import LinkedEditingRangeManager from './provider/linkedEditingRangeManager'
-import snippetManager from './snippets/manager'
-import sources from './sources'
-import { CompleteOption, CompleteResult, ISource, SourceType, ExtendedCompleteItem } from './types'
-import DiagnosticCollection from './diagnostic/collection'
-import * as complete from './util/complete'
-import { getChangedFromEdits, rangeOverlap } from './util/position'
-import { byteIndex, byteLength, byteSlice } from './util/string'
-import window from './window'
-import workspace from './workspace'
 const logger = require('./util/logger')('languages')
-
-export interface CompletionSource {
-  id: string
-  source: ISource
-  languageIds: string[]
-}
-
-interface CompleteConfig {
-  snippetsSupport: boolean
-  defaultKindText: string
-  priority: number
-  echodocSupport: boolean
-  detailMaxLength: number
-  detailField: string
-  invalidInsertCharacters: string[]
-  floatEnable: boolean
-}
 
 export interface DocumentSymbolProviderMetadata {
   /**
@@ -66,7 +39,6 @@ export interface DocumentSymbolProviderMetadata {
 class Languages {
   private _onDidSemanticTokensRefresh = new Emitter<DocumentSelector>()
   public readonly onDidSemanticTokensRefresh: Event<DocumentSelector> = this._onDidSemanticTokensRefresh.event
-  private completeConfig: CompleteConfig
   private onTypeFormatManager = new OnTypeFormatManager()
   private documentLinkManager = new DocumentLinkManager()
   private documentColorManager = new DocumentColorManager()
@@ -92,71 +64,6 @@ class Languages {
   private semanticTokensRangeManager = new SemanticTokensRangeManager()
   private linkedEditingManager = new LinkedEditingRangeManager()
   private cancelTokenSource: CancellationTokenSource = new CancellationTokenSource()
-  private completionItemKindMap: Map<CompletionItemKind, string>
-
-  public init(): void {
-    this.loadCompleteConfig()
-    workspace.onDidChangeConfiguration(e => {
-      if (e.affectsConfiguration('suggest')) {
-        this.loadCompleteConfig()
-      }
-    }, this)
-  }
-
-  private get nvim(): Neovim {
-    return workspace.nvim
-  }
-
-  private get detailField(): string {
-    let { detailField, floatEnable } = this.completeConfig
-    if (detailField == 'preview' && (!floatEnable || !workspace.floatSupported)) {
-      return 'menu'
-    }
-    return 'preview'
-  }
-
-  private loadCompleteConfig(): void {
-    let suggest = workspace.getConfiguration('suggest')
-    let labels = suggest.get<{ [key: string]: string }>('completionItemKindLabels', {})
-    this.completionItemKindMap = new Map([
-      [CompletionItemKind.Text, labels['text'] || 'v'],
-      [CompletionItemKind.Method, labels['method'] || 'f'],
-      [CompletionItemKind.Function, labels['function'] || 'f'],
-      [CompletionItemKind.Constructor, typeof labels['constructor'] == 'function' ? 'f' : labels['con' + 'structor']],
-      [CompletionItemKind.Field, labels['field'] || 'm'],
-      [CompletionItemKind.Variable, labels['variable'] || 'v'],
-      [CompletionItemKind.Class, labels['class'] || 'C'],
-      [CompletionItemKind.Interface, labels['interface'] || 'I'],
-      [CompletionItemKind.Module, labels['module'] || 'M'],
-      [CompletionItemKind.Property, labels['property'] || 'm'],
-      [CompletionItemKind.Unit, labels['unit'] || 'U'],
-      [CompletionItemKind.Value, labels['value'] || 'v'],
-      [CompletionItemKind.Enum, labels['enum'] || 'E'],
-      [CompletionItemKind.Keyword, labels['keyword'] || 'k'],
-      [CompletionItemKind.Snippet, labels['snippet'] || 'S'],
-      [CompletionItemKind.Color, labels['color'] || 'v'],
-      [CompletionItemKind.File, labels['file'] || 'F'],
-      [CompletionItemKind.Reference, labels['reference'] || 'r'],
-      [CompletionItemKind.Folder, labels['folder'] || 'F'],
-      [CompletionItemKind.EnumMember, labels['enumMember'] || 'm'],
-      [CompletionItemKind.Constant, labels['constant'] || 'v'],
-      [CompletionItemKind.Struct, labels['struct'] || 'S'],
-      [CompletionItemKind.Event, labels['event'] || 'E'],
-      [CompletionItemKind.Operator, labels['operator'] || 'O'],
-      [CompletionItemKind.TypeParameter, labels['typeParameter'] || 'T'],
-    ])
-    // let useFloat = workspace.floatSupported && suggest.get
-    this.completeConfig = {
-      defaultKindText: labels['default'] || '',
-      priority: suggest.get<number>('languageSourcePriority', 99),
-      echodocSupport: suggest.get<boolean>('echodocSupport', false),
-      snippetsSupport: suggest.get<boolean>('snippetsSupport', true),
-      detailField: suggest.get<string>('detailField', 'preview'),
-      detailMaxLength: suggest.get<number>('detailMaxLength', 100),
-      floatEnable: suggest.get<boolean>('floatEnable', true),
-      invalidInsertCharacters: suggest.get<string[]>('invalidInsertCharacters', ['(', '<', '{', '[', '\r', '\n']),
-    }
-  }
 
   public hasFormatProvider(doc: TextDocument): boolean {
     if (this.formatManager.hasProvider(doc)) {
@@ -185,15 +92,9 @@ class Languages {
     priority?: number,
     allCommitCharacters?: string[]
   ): Disposable {
-    selector = typeof selector == 'string' ? [selector] : selector
-    let source = this.createCompleteSource(name, shortcut, provider, selector, triggerCharacters, allCommitCharacters || [], priority)
-    sources.addSource(source)
-    logger.debug('created service source', name)
-    return {
-      dispose: () => {
-        sources.removeSource(name)
-      }
-    }
+    selector = typeof selector == 'string' ? [{ language: selector }] : selector
+    let sources = require('./sources/index').default
+    return sources.createLanguageSource(name, shortcut, selector, provider, triggerCharacters, priority, allCommitCharacters)
   }
 
   public registerCodeActionProvider(selector: DocumentSelector, provider: CodeActionProvider, clientId: string | undefined, codeActionKinds?: CodeActionKind[]): Disposable {
@@ -536,328 +437,13 @@ class Languages {
         throw new Error(`Invalid provider name: ${id}`)
     }
   }
-
-  public dispose(): void {
-    // noop
-  }
-
   public createDiagnosticCollection(owner: string): DiagnosticCollection {
     return diagnosticManager.create(owner)
-  }
-
-  private createCompleteSource(
-    name: string,
-    shortcut: string,
-    provider: CompletionItemProvider,
-    selector: DocumentSelector,
-    triggerCharacters: string[],
-    allCommitCharacters: string[],
-    priority?: number | undefined
-  ): ISource {
-    let filetype: string
-    // track them for resolve
-    let completeItems: CompletionItem[] = []
-    // line used for TextEdit
-    let hasResolve = typeof provider.resolveCompletionItem === 'function'
-    priority = priority == null ? this.completeConfig.priority : priority
-    // index set of resolved items
-    let resolvedIndexes: Set<number> = new Set()
-    let source: ISource = {
-      name,
-      priority,
-      shortcut,
-      enable: true,
-      sourceType: SourceType.Service,
-      documentSelector: selector,
-      triggerCharacters: triggerCharacters || [],
-      toggle: () => {
-        source.enable = !source.enable
-      },
-      doComplete: async (opt: CompleteOption, token: CancellationToken): Promise<CompleteResult | null> => {
-        let { triggerCharacter, bufnr } = opt
-        filetype = opt.filetype
-        resolvedIndexes = new Set()
-        let isTrigger = triggerCharacters && triggerCharacters.includes(triggerCharacter)
-        let triggerKind: CompletionTriggerKind = CompletionTriggerKind.Invoked
-        if (opt.triggerForInComplete) {
-          triggerKind = CompletionTriggerKind.TriggerForIncompleteCompletions
-        } else if (isTrigger) {
-          triggerKind = CompletionTriggerKind.TriggerCharacter
-        }
-        if (token.isCancellationRequested) return null
-        let position = complete.getPosition(opt)
-        let context: any = { triggerKind, option: opt }
-        if (isTrigger) context.triggerCharacter = triggerCharacter
-        let result
-        try {
-          let doc = workspace.getDocument(bufnr)
-          result = await Promise.resolve(provider.provideCompletionItems(doc.textDocument, position, token, context))
-        } catch (e) {
-          // don't disturb user
-          logger.error(`Complete "${name}" error:`, e)
-          return null
-        }
-        if (!result || token.isCancellationRequested) return null
-        completeItems = Array.isArray(result) ? result : result.items
-        if (!completeItems || completeItems.length == 0) return null
-        let startcol = this.getStartColumn(opt.line, completeItems)
-        let option: CompleteOption = Object.assign({}, opt)
-        let prefix: string
-        if (startcol != null) {
-          if (startcol < option.col) {
-            prefix = byteSlice(opt.line, startcol, option.col)
-          }
-          option.col = startcol
-        }
-        let items: ExtendedCompleteItem[] = completeItems.map((o, index) => {
-          let item = this.convertVimCompleteItem(o, shortcut, option, prefix)
-          item.index = index
-          return item
-        })
-        return {
-          startcol,
-          isIncomplete: !!(result as CompletionList).isIncomplete,
-          items
-        }
-      },
-      onCompleteResolve: async (item: ExtendedCompleteItem, token: CancellationToken): Promise<void> => {
-        let { index } = item
-        let resolving = completeItems[index]
-        if (!resolving || resolvedIndexes.has(index)) return
-        if (hasResolve) {
-          token.onCancellationRequested(() => {
-            resolvedIndexes.delete(index)
-          })
-          resolvedIndexes.add(index)
-          try {
-            let resolved = await Promise.resolve(provider.resolveCompletionItem(Object.assign({}, resolving), token))
-            if (token.isCancellationRequested) return
-            if (resolved) {
-              Object.assign(resolving, resolved)
-            } else {
-              resolvedIndexes.delete(index)
-            }
-          } catch (e) {
-            resolvedIndexes.delete(index)
-            logger.error(`Error on complete resolve: ${e.message}`, e.stack)
-          }
-        }
-        if (item.documentation == null) {
-          let { documentation, detail } = resolving
-          if (!documentation && !detail) return
-          let docs = []
-          if (detail && !item.detailShown && detail != item.word) {
-            detail = detail.replace(/\n\s*/g, ' ')
-            if (detail.length) {
-              let isText = /^[\w-\s.,\t]+$/.test(detail)
-              docs.push({ filetype: isText ? 'txt' : filetype, content: detail })
-            }
-          }
-          if (documentation) {
-            if (typeof documentation == 'string') {
-              docs.push({
-                filetype: 'markdown',
-                content: documentation
-              })
-            } else if (documentation.value) {
-              docs.push({
-                filetype: documentation.kind == 'markdown' ? 'markdown' : 'txt',
-                content: documentation.value
-              })
-            }
-          }
-          item.documentation = docs
-        }
-      },
-      onCompleteDone: async (vimItem: ExtendedCompleteItem, opt: CompleteOption): Promise<void> => {
-        let item = completeItems[vimItem.index]
-        if (!item) return
-        let line = opt.linenr - 1
-        if (item.insertText != null && !item.textEdit) {
-          item.textEdit = {
-            range: Range.create(line, opt.col, line, opt.colnr - 1),
-            newText: item.insertText
-          }
-        }
-        if (vimItem.line) Object.assign(opt, { line: vimItem.line })
-        try {
-          let isSnippet = await this.applyTextEdit(item, vimItem.word, opt)
-          let { additionalTextEdits } = item
-          if (additionalTextEdits && item.textEdit) {
-            let r = InsertReplaceEdit.is(item.textEdit) ? item.textEdit.replace : item.textEdit.range
-            additionalTextEdits = additionalTextEdits.filter(edit => {
-              let er = InsertReplaceEdit.is(edit) ? edit.replace : edit.range
-              if (rangeOverlap(r, er)) {
-                logger.error('Filtered overlap additionalTextEdit:', edit)
-                return false
-              }
-              return true
-            })
-          }
-          await this.applyAdditionalEdits(additionalTextEdits, opt.bufnr, isSnippet)
-          if (isSnippet) await snippetManager.selectCurrentPlaceholder()
-          if (item.command && commands.has(item.command.command)) {
-            void commands.execute(item.command)
-          }
-        } catch (e) {
-          logger.error('Error on CompleteDone:', e)
-        }
-      },
-      shouldCommit: (item: ExtendedCompleteItem, character: string): boolean => {
-        let completeItem = completeItems[item.index]
-        if (!completeItem) return false
-        let commitCharacters = completeItem.commitCharacters || allCommitCharacters
-        return commitCharacters.includes(character)
-      }
-    }
-    return source
   }
 
   private get token(): CancellationToken {
     this.cancelTokenSource = new CancellationTokenSource()
     return this.cancelTokenSource.token
-  }
-
-  private async applyTextEdit(item: CompletionItem, word: string, option: CompleteOption): Promise<boolean> {
-    let { nvim } = this
-    let { textEdit } = item
-    if (!textEdit) return false
-    let { line, bufnr, linenr } = option
-    let doc = workspace.getDocument(bufnr)
-    if (!doc) return false
-    let newText = textEdit.newText
-    let range = InsertReplaceEdit.is(textEdit) ? textEdit.replace : textEdit.range
-    let isSnippet = item.insertTextFormat === InsertTextFormat.Snippet
-    // replace inserted word
-    let start = line.substr(0, range.start.character)
-    let end = line.substr(range.end.character)
-    if (isSnippet && this.completeConfig.snippetsSupport === false) {
-      // could be wrong, but maybe best we can do.
-      isSnippet = false
-      newText = word
-    }
-    if (isSnippet) {
-      let currline = doc.getline(linenr - 1)
-      let endCharacter = currline.length - end.length
-      let r = Range.create(linenr - 1, range.start.character, linenr - 1, endCharacter)
-      // can't select, since additionalTextEdits would break selection
-      return await snippetManager.insertSnippet(newText, false, r, item.insertTextMode)
-    }
-    let newLines = `${start}${newText}${end}`.split(/\r?\n/)
-    if (newLines.length == 1) {
-      await nvim.call('coc#util#setline', [linenr, newLines[0]])
-      await window.moveTo(Position.create(linenr - 1, (start + newText).length))
-    } else {
-      let buffer = nvim.createBuffer(bufnr)
-      await buffer.setLines(newLines, {
-        start: linenr - 1,
-        end: linenr,
-        strictIndexing: false
-      })
-      let line = linenr - 1 + newLines.length - 1
-      let character = newLines[newLines.length - 1].length - end.length
-      await window.moveTo({ line, character })
-    }
-    return false
-  }
-
-  private async applyAdditionalEdits(
-    textEdits: TextEdit[],
-    bufnr: number,
-    snippet: boolean): Promise<void> {
-    if (!textEdits || textEdits.length == 0) return
-    let document = workspace.getDocument(bufnr)
-    if (!document) return
-    await document.patchChange(true)
-    // move cursor after edit
-    let changed = null
-    let pos = await window.getCursorPosition()
-    if (!snippet) changed = getChangedFromEdits(pos, textEdits)
-    await document.applyEdits(textEdits)
-    if (changed) await window.moveTo(Position.create(pos.line + changed.line, pos.character + changed.character))
-  }
-
-  private getStartColumn(line: string, items: CompletionItem[]): number | null {
-    let first = items[0]
-    if (!first.textEdit) return null
-    let { newText } = first.textEdit
-    let range = InsertReplaceEdit.is(first.textEdit) ? first.textEdit.replace : first.textEdit.range
-    let { character } = range.start
-    if (newText.length < range.end.character - character) {
-      return null
-    }
-    for (let i = 0; i < 10; i++) {
-      let o = items[i]
-      if (!o) break
-      if (!o.textEdit) return null
-      if (InsertReplaceEdit.is(o.textEdit)) return null
-      let r = InsertReplaceEdit.is(o.textEdit) ? o.textEdit.replace : o.textEdit.range
-      if (r.start.character !== character) return null
-    }
-    return byteIndex(line, character)
-  }
-
-  private convertVimCompleteItem(item: CompletionItem, shortcut: string, opt: CompleteOption, prefix: string): ExtendedCompleteItem {
-    let { echodocSupport, detailMaxLength, invalidInsertCharacters } = this.completeConfig
-    let { detailField } = this
-    let hasAdditionalEdit = item.additionalTextEdits && item.additionalTextEdits.length > 0
-    let isSnippet = item.insertTextFormat === InsertTextFormat.Snippet || hasAdditionalEdit
-    let label = item.label.trim()
-    let obj: ExtendedCompleteItem = {
-      word: complete.getWord(item, opt, invalidInsertCharacters),
-      abbr: label,
-      menu: `[${shortcut}]`,
-      kind: complete.completionKindString(item.kind, this.completionItemKindMap, this.completeConfig.defaultKindText),
-      sortText: item.sortText || null,
-      sourceScore: item['score'] || null,
-      filterText: item.filterText || label,
-      isSnippet,
-      dup: item.data && item.data.dup == 0 ? 0 : 1
-    }
-    if (prefix) {
-      if (!obj.filterText.startsWith(prefix)) {
-        if (item.textEdit && item.textEdit.newText.startsWith(prefix)) {
-          obj.filterText = item.textEdit.newText.split(/\r?\n/)[0]
-        }
-      }
-      if (!item.textEdit && !obj.word.startsWith(prefix)) {
-        // fix remains completeItem that should not change startcol
-        obj.word = `${prefix}${obj.word}`
-      }
-    }
-    if (item && item.detail && detailField != 'preview') {
-      let detail = item.detail.replace(/\n\s*/g, ' ')
-      if (byteLength(detail) < detailMaxLength) {
-        if (detailField == 'menu') {
-          obj.menu = `${detail} ${obj.menu}`
-        } else if (detailField == 'abbr') {
-          obj.abbr = `${obj.abbr} - ${detail}`
-        }
-        obj.detailShown = 1
-      }
-    }
-    if (item.documentation) {
-      obj.info = typeof item.documentation == 'string' ? item.documentation : item.documentation.value
-    } else {
-      obj.info = ''
-    }
-    if (obj.word == '') obj.empty = 1
-    if (item.textEdit) obj.line = opt.line
-    if (item.kind == CompletionItemKind.Folder && !obj.abbr.endsWith('/')) {
-      obj.abbr = obj.abbr + '/'
-    }
-    if (echodocSupport && item.kind >= 2 && item.kind <= 4) {
-      let fields = [item.detail || '', obj.abbr, obj.word]
-      for (let s of fields) {
-        if (s.includes('(')) {
-          obj.signature = s
-          break
-        }
-      }
-    }
-    if (item.preselect) obj.preselect = true
-    if (item.data?.optional) obj.abbr = obj.abbr + '?'
-    return obj
   }
 }
